+++ src/main/java/org/c4marathon/assignment/application/AccountService.java
@@ -5,7 +5,7 @@ import static org.c4marathon.assignment.global.CommonUtils.*;
 
 import org.c4marathon.assignment.domain.Account;
 import org.c4marathon.assignment.domain.AccountRepository;
-import org.c4marathon.assignment.domain.AccountType;
+import org.c4marathon.assignment.domain.type.AccountType;
 import org.c4marathon.assignment.domain.UserRepository;
 import org.c4marathon.assignment.domain.dto.response.CreatedAccountInfo;
 import org.c4marathon.assignment.domain.dto.response.TransferResult;
