+++ src/main/java/org/c4marathon/assignment/application/TransactionService.java
@@ -1,0 +1,7 @@
+package org.c4marathon.assignment.application;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class TransactionService {
+}
