--- src/main/java/org/c4marathon/assignment/domain/SendingType.java
@@ -1,6 +1,0 @@
-package org.c4marathon.assignment.domain;
-
-public enum SendingType {
-	EAGER,
-	LAZY
-}
