+++ src/main/java/org/c4marathon/assignment/domain/dto/request/SettlementRequest.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.domain.dto.request;
 
 import java.util.List;
 
-import org.c4marathon.assignment.domain.SettlementType;
+import org.c4marathon.assignment.domain.type.SettlementType;
 
 public record SettlementRequest(List<Long> userIds, long money, SettlementType settlementType) {
 }
