+++ src/main/java/org/c4marathon/assignment/domain/dto/response/CreatedAccountInfo.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.domain.dto.response;
 
 import java.time.LocalDateTime;
 
-import org.c4marathon.assignment.domain.AccountType;
+import org.c4marathon.assignment.domain.type.AccountType;
 
 public record CreatedAccountInfo(String accountNumber, LocalDateTime createdAt, AccountType accountType, long balance,
 								 boolean isMain) {
