+++ src/main/java/org/c4marathon/assignment/domain/type/AccountType.java
@@ -1,4 +1,4 @@
-package org.c4marathon.assignment.domain;
+package org.c4marathon.assignment.domain.type;
 
 public enum AccountType {
 	CHECKING, // 입출금(일반)
