+++ src/main/java/org/c4marathon/assignment/domain/type/SendingType.java
@@ -1,0 +1,6 @@
+package org.c4marathon.assignment.domain.type;
+
+public enum SendingType {
+	EAGER,
+	LAZY
+}
